@@ -16,7 +16,6 @@ package actions
 
 import (
 	"github.com/jptosso/coraza-waf/pkg/engine"
-	"github.com/jptosso/coraza-waf/pkg/lua"
 	"github.com/jptosso/coraza-waf/pkg/utils"
 )
 
@@ -34,9 +33,7 @@ func (a *Exec) Init(r *engine.Rule, data string) string {
 }
 
 func (a *Exec) Evaluate(r *engine.Rule, tx *engine.Transaction) {
-	l := &lua.LuaScript{}
-	l.FromString(a.cachedScript)
-	l.Evaluate(tx, 1000)
+	// Not implemented
 }
 
 func (a *Exec) GetType() int {
