@@ -66,12 +66,12 @@ func (c *PersistentCollection) New(engine PersistenceEngine, webappid string, co
 	c.webapp = webappid
 	timenow := strconv.FormatInt(time.Now().UnixNano(), 10)
 	c.data = map[string][]string{
-		"CREATE_TIME":      []string{timenow},
-		"IS_NEW":           []string{"1"},
-		"LAST_UPDATE_TIME": []string{timenow},
-		"TIMEOUT":          []string{"0"},
-		"UPDATE_COUNTER":   []string{"0"},
-		"UPDATE_RATE":      []string{"0"},
+		"CREATE_TIME":      {timenow},
+		"IS_NEW":           {"1"},
+		"LAST_UPDATE_TIME": {timenow},
+		"TIMEOUT":          {"0"},
+		"UPDATE_COUNTER":   {"0"},
+		"UPDATE_RATE":      {"0"},
 	}
 }
 
