@@ -17,9 +17,6 @@ package engine
 import (
 	"bufio"
 	"fmt"
-	"github.com/antchfx/xmlquery"
-	"github.com/jptosso/coraza-waf/pkg/utils"
-	log "github.com/sirupsen/logrus"
 	"html"
 	"io"
 	"net/http"
@@ -29,6 +26,11 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/antchfx/xmlquery"
+	"github.com/jptosso/coraza-waf/pkg/engine/loggers"
+	"github.com/jptosso/coraza-waf/pkg/utils"
+	log "github.com/sirupsen/logrus"
 )
 
 type Interruption struct {
@@ -157,14 +159,12 @@ func (tx *Transaction) Init(waf *Waf) error {
 	tx.ResponseBodyLimit = 524288
 	tx.ResponseBodyMimeType = []string{"text/html", "text/plain"}
 	tx.RuleEngine = tx.Waf.RuleEngine
-	tx.AuditLogType = tx.Waf.AuditLogType
 	tx.Skip = 0
 	tx.PersistentCollections = map[string]*PersistentCollection{}
 	tx.RuleRemoveTargetById = map[int][]*KeyValue{}
 	tx.RuleRemoveById = []int{}
 	tx.StopWatches = map[int]int{}
 	tx.RequestBodyReader = NewBodyReader(tx.Waf.TmpDir, tx.Waf.RequestBodyInMemoryLimit)
-	// TODO add response values
 	tx.ResponseBodyReader = NewBodyReader(tx.Waf.TmpDir, tx.Waf.RequestBodyInMemoryLimit)
 
 	return nil
@@ -434,17 +434,21 @@ func (tx *Transaction) GetRemovedTargets(id int) []*KeyValue {
 
 func (tx *Transaction) ToAuditJson() []byte {
 	al := tx.ToAuditLog()
-	return al.ToJson()
+	data, _ := al.JSON()
+	return data
 }
 
-func (tx *Transaction) ToAuditLog() *AuditLog {
-	al := &AuditLog{}
-	al.Init(tx)
+func (tx *Transaction) ToAuditLog() *loggers.AuditLog {
+	al := &loggers.AuditLog{}
 	return al
 }
 
 func (tx *Transaction) saveLog() error {
-	return tx.Waf.Logger.WriteAudit(tx)
+	for _, l := range tx.Waf.Loggers() {
+		l.Write(tx.ToAuditLog())
+	}
+
+	return nil
 }
 
 // Save persistent collections to persistence engine
@@ -815,4 +819,85 @@ func (tx *Transaction) ProcessLogging() {
 		return
 	}
 	tx.saveLog()
+}
+
+// AuditLog returns an AuditLog struct
+func (tx *Transaction) AuditLog() *loggers.AuditLog {
+	al := &loggers.AuditLog{}
+	parts := tx.AuditLogParts
+	al.Messages = []*loggers.AuditMessage{}
+	ts := time.Unix(0, tx.Timestamp).Format("02/Jan/2006:15:04:20 -0700")
+	al.Transaction = &loggers.AuditTransaction{
+		Timestamp:  ts,
+		Id:         tx.Id,
+		ClientIp:   tx.GetCollection(VARIABLE_REMOTE_ADDR).GetFirstString(""),
+		ClientPort: tx.GetCollection(VARIABLE_REMOTE_PORT).GetFirstInt(""),
+		HostIp:     "",
+		HostPort:   0,
+		ServerId:   "",
+		Request: &loggers.AuditTransactionRequest{
+			Protocol:    tx.GetCollection(VARIABLE_REQUEST_METHOD).GetFirstString(""),
+			Uri:         tx.GetCollection(VARIABLE_REQUEST_URI).GetFirstString(""),
+			HttpVersion: tx.GetCollection(VARIABLE_REQUEST_PROTOCOL).GetFirstString(""),
+			//Body and headers are audit parts
+		},
+		Response: &loggers.AuditTransactionResponse{
+			Status: tx.GetCollection(VARIABLE_RESPONSE_STATUS).GetFirstInt(""),
+			//body and headers are audit parts
+		},
+	}
+
+	for _, p := range parts {
+		switch p {
+		case 'B':
+			al.Transaction.Request.Headers = tx.GetCollection(VARIABLE_REQUEST_HEADERS).GetData()
+		case 'C':
+			al.Transaction.Request.Body = tx.GetCollection(VARIABLE_REQUEST_BODY).GetFirstString("")
+		case 'F':
+			al.Transaction.Response.Headers = tx.GetCollection(VARIABLE_RESPONSE_HEADERS).GetData()
+		case 'G':
+			al.Transaction.Response.Body = tx.GetCollection(VARIABLE_RESPONSE_BODY).GetFirstString("")
+		case 'H':
+			servera := tx.GetCollection(VARIABLE_RESPONSE_HEADERS).Get("server")
+			server := ""
+			if len(server) > 0 {
+				server = servera[0]
+			}
+			al.Transaction.Producer = &loggers.AuditTransactionProducer{
+				Connector:  "unknown",
+				Version:    "unknown",
+				Server:     server,
+				RuleEngine: tx.RuleEngine,
+				Stopwatch:  tx.GetStopWatch(),
+			}
+		case 'I':
+			// not implemented
+			// TODO
+		case 'J':
+			//upload data
+			// TODO
+		case 'K':
+			for _, mr := range tx.MatchedRules {
+				r := mr.Rule
+				al.Messages = append(al.Messages, &loggers.AuditMessage{
+					Actionset: "",
+					Message:   "",
+					Data: &loggers.AuditMessageData{
+						File: "",
+						Line: 0,
+						Id:   r.Id,
+						Rev:  r.Rev,
+						Msg:  tx.MacroExpansion(r.Msg),
+						Data: "",
+						//Severity: r.Severity,
+						//Ver: r.Ver,
+						//Maturity: r.Maturity,
+						//Accuracy: r.Accuracy,
+						Tags: r.Tags,
+					},
+				})
+			}
+		}
+	}
+	return al
 }
