@@ -16,23 +16,18 @@ package operators
 
 import (
 	"github.com/jptosso/coraza-waf/pkg/engine"
-	_ "strings"
+	"strings"
 )
 
 type BeginsWith struct {
 	data string
-	dlen int
 }
 
 func (o *BeginsWith) Init(data string) {
 	o.data = data
-	o.dlen = len(data)
 }
 
 func (o *BeginsWith) Evaluate(tx *engine.Transaction, value string) bool {
-	if len(value) < o.dlen {
-		return false
-	}
-	return o.data == value[0:o.dlen]
-	//return strings.HasPrefix(value, o.data)
+	data := tx.MacroExpansion(o.data)
+	return strings.HasPrefix(value, data)
 }
